Update project code

+++ src/main/java/com/example/demodatn2/dto/GenerateQuickVariantsRequest.java
@@ -1,6 +1,7 @@
 package com.example.demodatn2.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -8,7 +9,7 @@ public class GenerateQuickVariantsRequest {
     private List<String> colors;        // ["Đen", "Trắng", ...]
     private List<String> sizes;         // ["XS", "S", "M", ...]
     private Integer stock;              // số lượng tồn mặc định
-    private Integer price;              // giá bán mặc định
+    private BigDecimal price;           // giá bán mặc định
     private Integer weight;             // khối lượng (gram)
     private String mode;                // "replace_all" hoặc "add_missing"
     private String maSanPham;           // mã sản phẩm để sinh SKU
